pyproject.toml
[tool.poetry]
name = "unidoc-service"
version = "0.1.0"
description = ""
authors = ["Aline Rose <aline.rose@mail.uft.edu.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.103.1"
uvicorn = "0.23.2"
pypdf2 = "3.0.1"
spacy = "3.6.1"
requests = "2.31.0"
gensim = "^4.3.2"
sqlalchemy = "2.0.23"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
jwt = "^1.3.1"
pyjwt = "^2.8.0"
docarray = "0.21"
nltk = "^3.8.1"
scikit-learn = "^1.4.2"
pandas = "^2.2.2"
alembic = "^1.13.1"
prettytable = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
